move shared test funcs out of monad-consensus into a monad-testutil. Refactor unit tests in monad-consensus integration tests as necessary
