+++ Cargo.toml
@@ -1,10 +1,11 @@
 [workspace]
 
 members = [
+    "monad-blocktree",
     "monad-consensus",
     "monad-crypto",
     "monad-executor",
     "monad-state",
+    "monad-testutil",
     "monad-validator",
-    "monad-blocktree",
 ]
