+++ monad-consensus/Cargo.toml
@@ -11,3 +11,6 @@ monad-crypto = { path = "../monad-crypto" }
 sha2 = "0.10.6"
 test-case = "3.0.0"
 zerocopy = "0.6.1"
+
+[dev-dependencies]
+monad-testutil = { path = "../monad-testutil"}
