+++ monad-consensus/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod mock_types;
 pub mod signatures;
 pub mod types;
 pub mod validation;
