--- monad-consensus/src/mock_types/mod.rs
@@ -1,1 +1,0 @@
-pub mod mock_signature;
