+++ monad-consensus/src/types/quorum_certificate.rs
@@ -4,7 +4,7 @@ use crate::*;
 use super::{ledger::*, signature::ConsensusSignature, voting::*};
 
 #[non_exhaustive]
-#[derive(Clone, Debug)]
+#[derive(Clone, Default, Debug)]
 pub struct QuorumCertificate<T>
 where
     T: VotingQuorum,
@@ -72,7 +72,7 @@ pub struct QcInfo {
 }
 
 #[derive(Copy, Clone, Debug)]
-pub struct Rank(QcInfo);
+pub struct Rank(pub QcInfo);
 
 impl PartialEq for Rank {
     fn eq(&self, other: &Self) -> bool {
@@ -102,51 +102,5 @@ impl<T: VotingQuorum> QuorumCertificate<T> {
             signatures,
             signature_hash: hash,
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::mock_types::mock_signature::MockSignatures;
-    use crate::types::ledger::LedgerCommitInfo;
-    use crate::types::voting::VoteInfo;
-
-    use crate::*;
-
-    use super::QcInfo;
-    use super::QuorumCertificate;
-    use super::Rank;
-
-    #[test]
-    fn comparison() {
-        let ci = LedgerCommitInfo::default();
-
-        let mut vi_1 = VoteInfo::default();
-        vi_1.round = Round(2);
-
-        let mut vi_2 = VoteInfo::default();
-        vi_2.round = Round(3);
-
-        let qc_1 = QuorumCertificate::<MockSignatures>::new(
-            QcInfo {
-                vote: vi_1,
-                ledger_commit: ci,
-            },
-            MockSignatures(),
-        );
-        let mut qc_2 = QuorumCertificate::<MockSignatures>::new(
-            QcInfo {
-                vote: vi_2,
-                ledger_commit: ci,
-            },
-            MockSignatures(),
-        );
-
-        assert!(Rank(qc_1.info) < Rank(qc_2.info));
-        assert!(Rank(qc_2.info) > Rank(qc_1.info));
-
-        qc_2.info.vote.round = Round(2);
-
-        assert_eq!(Rank(qc_1.info), Rank(qc_2.info));
     }
 }
