+++ monad-testutil/Cargo.toml
@@ -1,13 +1,10 @@
 [package]
-name = "monad-blocktree"
+name = "monad-testutil"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-monad-consensus = { path = "../monad-consensus" }
-ptree = "0.4.0"
-
-[dev-dependencies]
-monad-testutil = { path = "../monad-testutil" }
+monad-consensus = {path = "../monad-consensus"}
+sha2 = "0.10.6"
